Address code review feedback for remote symbolication
This commit addresses code review findings and adds observability improvements:

1. **Fix type mismatch in getLibraryByIndex**
   - Changed parameter from uint16_t to uint32_t
   - Matches lib_index packing width (17 bits = max 131K libraries)
   - Prevents theoretical overflow for large library counts

2. **Clarify library array stability during serialization**
   - Added comment explaining lockAll() ensures array stability
   - Corrected initial misunderstanding about race conditions
   - No actual race exists - serialization happens with profiling locked

3. **Add metrics for remote symbolication**
   - REMOTE_SYMBOLICATION_FRAMES: Track frames using remote symbolication
   - REMOTE_SYMBOLICATION_LIBS_WITH_BUILD_ID: Libraries with extracted build-IDs
   - REMOTE_SYMBOLICATION_BUILD_ID_CACHE_HITS: Cache efficiency tracking
   - Enables production monitoring of feature adoption and performance

All changes verified with RemoteSymbolicationTest passing.

+++ ddprof-lib/src/main/cpp/libraries.h
@@ -21,7 +21,8 @@ class Libraries {
   CodeCache *findLibraryByAddress(const void *address);
 
   // Get library by index (used for remote symbolication unpacking)
-  CodeCache *getLibraryByIndex(uint16_t index) const {
+  // Note: Parameter is uint32_t to match lib_index packing (17 bits = max 131K libraries)
+  CodeCache *getLibraryByIndex(uint32_t index) const {
     if (index < _native_libs.count()) {
       return _native_libs[index];
     }
